% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2p.R
\name{fp_parse_config}
\alias{fp_parse_config}
\title{Parse a TOML confifugration file}
\usage{
fp_parse_config(config_file)
}
\arguments{
\item{config_file}{Path to the TOML config file.}
}
\description{
Parse a TOML confifugration file
}
