% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl.R
\name{tpl_replace_whisker}
\alias{tpl_replace_whisker}
\title{Replaces whiskers in a template string}
\usage{
tpl_replace_whisker(
  txt,
  values,
  open = "{{",
  close = "}}",
  require_all_value = FALSE
)
}
\arguments{
\item{txt}{the tpl as a string}

\item{values}{named list of values}
}
\description{
Replaces whiskers in a template string
}
